requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==2.1
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.9.6
commontools==0.1.0
itsdangerous==0.24
pyserial==3.1.1